Created the infrastructure to add more region buttons to home page
* Created an enum to hold all region data with methods to get all region
data and get chart data lists
* Created a dashboard statistics class that will hold various statistics
that will eventually get populated at the top of the dashboard page
* Updated the dashboard controller to populate the dashboard statistics
and to use the enum instead of switch-case. This allows for any number
of regions without having a huge switch-case or if/else structure.
* Updated the home page jsp to have one more button (Germany) and to use
enum values for the ${<enum region>} for each button
* Updated the dashboard jsp to use the new dashboard object structure