Figured out how to get dates in tool tips on charts
* In dashboard.jsp, used a custom contentFormatter to format the tool
tips to include the date and the y value of each of the data series on
the chart.
* In ChartListServiceImpl, added a "dateChecked" field key to each of
the xyPair map objects that includes the dateChecked.toString() created
from the API data sources.
* In dashboard.jsp, when creating the dataPointsArr arrays for the two
data series on each chart, add an indexLabel field which the
contentFormatter places into the tool tip for the value next to "Date:
".
* In dashboard.jsp, set the indexLabelFontSize to zero so the labels do
not show next to the data points on the charts.
* This is a hacky solution, but achieves the effect of showing the date
of each data point in the tool tip, but not showing the indexLabels on
the chart.