Finished creating the ability to make a dashboard for a combined region
* Can click multiple states from the front-end and it will now make a
dashboard for the sum of all data from the selected states.
* took some data structure shenanigans to make it work. Found the latest
date of all the states for cases > 0 and same for deaths > 0. Then
iterated through each state's data list and put all relevant data in a
group of TreeMaps with key = integer date that each contain the sum of
the field values for all selected states. Then, iterated through the key
set and constructed new UnitedStatesData objects and populated them with
the values from the sum maps using each integer date key. This basically
aligns all the sums by date.