Another attempt at a fix for cache refresh - noncacheable method
* Cache for world data still not refreshing properly. I think it's
because the cache repopulation method was calling
makeDataListFromExternalSource, which itself was cacheable, so the
cacheput was calling a cacheable method which was, I think, just
returning was was in the cache.
* Solution (I hope): Leave makeDataListFromExternalSource as @Cacheable,
but move its guts to a non-cacheable method called
getDataFromWorldSource. Then, the @Cacheable method will call this new
method which Spring will intercept and retrieve from the cache, but when
the @CachePut repopulateCache method calls getDataFromWorldSource, it
will actually call the getDataFromWorldSource method and refresh the
data from the source. I think. I hope.