Added the ability to download the chart data from a dashboard - STUB
* Added a listener in the DashboardController for a POST request to
/download which calls a downloadChartData method.
* Updated the dashboard.jsp to use a Spring form for the Download Chart
Data button that sends a POST request to /download.
* Created a new DownloadChartData class that has a downloadChartData
method that will ultimately populate the CSV data and write it to the
OutputStream.
* Added the region string as a field in the DashboardChart class which
DashboardChartServiceImpl populates for each chart. I don't like thi
solution, but it was an easy way to get the region attached to the chart
data so it can ultimately be written in the CSV.
* In the DashboardController, added a field to hold the list of
DashboardChart objects so the list is available to the /download
endpoint when it gets called. This alleviates the need to somehow get
the dashboard chart data back from the JSP.