Fixed the Our World In Data source
* The OWID source went FROM key-value pairs where the key was a country
code and the value was an array of data objects containing all the data
about the country TO key-value pairs where the key is still the country
code, but the value is now a single object with all the base country
data k-v pairs with a "data" key containing an array of data objects.
* The keys in the data object array were the same as before.
* Had to refactor the ExternalDataServiceWorldImpl class to account for
the new JSON structure as well as update the
WorldDataSourceOurWorldInData class by moving the data objects to a new
class called WorldDataOWID
* Added V2 master json data files for reference in case the json files
change again
* Changed the .properties files to use OWID until the Euro CDC data is
fixed (may not even return to Euro data because it is weekly now).