Finished building new USA data getter as a multi-region service
* At application startup, gets data for all 50 states and stitches them
together and puts the final list in the cache for faster retrieval.
*When UnitedStates button is clicked, the program makes the normal call
to get "USA" data and the makeDataListFromExternalSource method pulls
the US data from the cache. If not present, it calls
multiRegionDataService.getMultiRegionDataFromExternalSource which gets
the data from all 50 states, stitches them and returns the list. This
should ONLY happen during application startup and cache eviction /
repopulation.