Fixed HTTP 500 errors when selecting a country
* The ExternalDataListGetterWorld was broken. It was not filtering the
entire world data list by the country selected. Added the filter back in
from the old ExternalDataServiceSingleCountryImpl class that was deleted
as part of the refactoring of external data list getters
(https://github.com/roychancellor/coronavirustracker/pull/38/files).
* The ExternalDataListGetterFactory was also broken, as it was not
assigning an object to cases where the region string was not USA but was
3 characters (the indicator of a non-USA country). Added the case back
in.
* Changed the name of ChartListConstants to DataTransformConstants and
added the threshold for adding a cases value to a list as an enum.
* Removed the US_CUTOFF_DATE constant from the IExternalDataListGetter
interface and put it as an enum in DataTransformConstants.